UI status chip with severity, src/light/layers/ subfolder, Filesystem+Scheduler split
KPI: 16384lights | PC:304KB | tick:48/51/148/147us(FPS:20833/19607/6756/6802) | ESP32:920KB | tick:54004us(FPS:18) | heap:185KB | src:52(7630) | test:30(3715) | lizard:24w

Bundled UI work, structural refactors, and the per-module status mechanism.

Status / severity slot on MoonModule
-------------------------------------

New `MoonModule::status() + severity()` slot. A short message any module
attaches when it wants the user to see something — Layer writes "buffer
reduced — not enough memory" on memory degradation, NetworkModule
writes "Eth: 192.168.1.210" or "No network". Severity is
Status / Warning / Error → UI picks ℹ️ / ⚠️ / ❌. Wire cost: +9 bytes
per module (pointer + enum), zero allocation (callers pass flash
literals or own a static buffer).

JSON wire (only emitted when set):
  "status":"Eth: 192.168.1.210","severity":"status"

NetworkModule drops its `status` ReadOnly control and routes through
setStatus() per state transition — same string, no separate control row.
Spec updated (NetworkModule.md).

src/light/layers/ subfolder
---------------------------

Moved Layer.h, Layers.h, Buffer.h, BlendMap.h, MappingLUT.h into
src/light/layers/. Symmetric to drivers/, effects/, layouts/,
modifiers/. ~34 includes updated mechanically.

FilesystemModule and Scheduler split to .h + .cpp
-------------------------------------------------

Two more core service modules migrated to .h + .cpp per
coding-standards.md. Added to mm_core's STATIC sources in both
CMakeLists.txt files. Removes the "lazy migration" entry from plan.md
— SystemModule and NetworkModule stay header-only.

UI: collapsible controls, drag rework
-------------------------------------

- Container modules wrap their own controls in a `<details>` element
  (default collapsed). Native HTML5 disclosure, no JS state.
- Drag rework: the whole card body is the drag source, controls
  (`.control-row`, `.card-controls-collapse > summary`) excluded.
  Drag handle (`☰`) glyph removed. Header buttons stay clickable
  (click vs drag is timing-based, not source-based).

  Implementation note: HTML5 `dragstart`'s `e.target` is always the
  draggable element (the card), not the deepest element under the
  mouse — `closest(".control-row")` never matches there. The reliable
  signal is the *mousedown* target, so we toggle `card.draggable` on
  mousedown based on where the grab landed. Touchstart mirror gives
  the same behaviour on mobile.

Two doc fence labels added (architecture.md, building.md — MD040).

Reviewer agent pass
-------------------

Opus reviewer ran over the staged diff (Event 1 gate 8, new policy
this commit). Findings addressed:

- NetworkModule.h stale comments (2 sites) — fixed: the
  pointer-bound-ReadOnly justification was rewritten to name the new
  MoonModule::status_ pointer it actually refers to.
- HttpServerModule duplicated `status/severity` emit block at two
  sites — extracted to a private static `writeStatus(sink, mod)`
  helper, shared by writeModuleJson (/api/state) and
  writeModuleMetricsJson (/api/system) so the two endpoints stay in
  sync.
- Reviewer flagged MirrorModifier::mirrorZ default change (false →
  true) as "buried in a refactor commit." Product owner confirmed the
  change is intended (3D layouts should mirror all three axes by
  default); calling it out explicitly here so it's not a surprise.

Findings accepted with reasons:

- JSON severity value "status" reads oddly next to the field name
  "status" (e.g. `"severity":"status"`). Kept — value mirrors the C++
  enum name (Status / Warning / Error → lowercased), HttpServerModule.md
  documents it; renaming to "info" would mean a 4-file cascade for a
  cosmetic improvement. Worth revisiting if a real reader is confused.
- Layer.md doesn't document the new status messages. Skipped — module
  specs are end-user wire contracts (per the principle in CLAUDE.md
  § Documentation), and these strings are visible in the UI chip
  already; not a control name so check_specs.py doesn't enforce.

Pre-commit gates: 1-7 all green (build / ctest / scenarios /
platform-boundary / specs / ESP32-default-build / KPI). The Reviewer
agent ran on the staged diff under the now-superseded "reviewer at
commit" policy that this commit *also* removes (the reviewer now runs
at PR-merge over the whole branch diff; mid-branch it's an on-demand
safety valve). Reviewer findings addressed above.
