+++ docs/moonmodules/core/NetworkModule.md
@@ -21,7 +21,7 @@ When a higher-priority connection becomes available, lower ones are torn down to
 - `addressing` (dropdown: DHCP / Static) — IP addressing mode (applies to both Ethernet and WiFi STA)
 - When Static: `ip`, `gateway`, `subnet` (text controls, shown dynamically via onBuildControls)
 - `dns` (text, optional) — DNS server. Empty = use gateway as DNS.
-- `status` (read-only) — current state: "Ethernet 192.168.1.210", "WiFi STA 192.168.1.100", "AP 4.3.2.1"
+No `status` *control*; the module surfaces its state via the generic `MoonModule::status()` slot — "Eth: 192.168.1.210", "WiFi: 10.0.0.5", "AP: MM-XXX @ 4.3.2.1", or "No network". The UI renders it as a chip in the card header (ℹ️ when connected, ❌ when no network) rather than a control row.
 
 Dynamic controls: when `addressing` changes, onBuildControls is called to show/hide the static IP fields.
