+++ esp32/main/CMakeLists.txt
@@ -3,6 +3,8 @@ idf_component_register(
         "main.cpp"
         "../../src/main.cpp"
         "../../src/core/HttpServerModule.cpp"
+        "../../src/core/FilesystemModule.cpp"
+        "../../src/core/Scheduler.cpp"
         "../../src/platform/esp32/platform_esp32.cpp"
     INCLUDE_DIRS
         "../../src"
