+++ src/light/layers/Layers.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "core/MoonModule.h"
-#include "light/Layer.h"
+#include "light/layers/Layer.h"
 #include "light/layouts/Layouts.h"
 #include "platform/platform.h"
