+++ test/test_layers_container.cpp
@@ -1,5 +1,5 @@
 #include "doctest.h"
-#include "light/Layers.h"
+#include "light/layers/Layers.h"
 #include "light/layouts/GridLayout.h"
 #include "light/effects/RainbowEffect.h"
 #include "light/effects/CheckerboardEffect.h"
