+++ test/test_preview_driver.cpp
@@ -1,6 +1,6 @@
 #include "doctest.h"
 #include "light/drivers/PreviewDriver.h"
-#include "light/Layer.h"
+#include "light/layers/Layer.h"
 #include "light/layouts/Layouts.h"
 #include "light/layouts/GridLayout.h"
 #include "core/PreviewFrame.h"
